refactored and moved regex function to helpers
